Refactor main into subcommands
